+++ project-griffin/project-griffin/source/utility/profile/Profile.h
@@ -30,24 +30,24 @@ namespace griffin {
 
 	/**
 	 * @class Profile
-	 * Profile is an RAII-type container meant to be placed on the stack and will time from creation
-	 * to destruction. Each time the profile is run
+	 * Profile is an RAII-type container meant to be placed on the stack and will time from
+	 * creation to destruction.
 	 *
 	 * Profile must be thread safe but should not introduce locks that increase contention of the
-	 * normally running program. For this reason, unique storage is required for each thread that is
-	 * profiled, including a separate stack of Profiles and a separate ProfileAggregate. A master index
-	 * of references to all thread-specific lists is kept. This can be problematic for thread pools
-	 * when the same task can be picked up by any worker thread at random. The aggregate information
-	 * for the task will be spread over several lists, when presumably you would want to see them
-	 * combined into one entry to capture the total aggregate. A post processing step could look at all
-	 * of the thread aggregates and recombine split profiles by name, and also indicate how many unique
-	 * threads the profile was taken on over its lifetime.
+	 * normally running program. For this reason, unique storage is required for each thread that
+	 * is profiled, including a separate stack of Profiles and a separate ProfileAggregate. A
+	 * master index of references to all thread-specific lists is kept. This can be problematic for
+	 * thread pools when the same task can be picked up by any worker thread at random. The
+	 * information for one task will be spread over several lists, but you would actually want to
+	 * see it combined into one aggregated entry. A post processing step could look at all of the
+	 * thread lists and combine split profiles by name, and also indicate how many unique threads
+	 * the profile was run on over its lifetime.
 	 *
 	 * std::this_thread::get_id().hash() is used to index each thread in the master map
 	 */
 	template <int Th>
 	struct Profile : auto_stacker<Profile<Th>> {
-		int32_t		m_frame;		//!< frame the code block is called on
+		int64_t		m_frame;		//!< frame the code block is called on
 		int64_t		m_startCounts;	//!< counts at start of profiling block
 		const char*	m_name;			//!< name of this profile
 		const char*	m_file;			//!< source file (from macro)
@@ -60,7 +60,7 @@ namespace griffin {
 		 * @param	file	source file of the profile block
 		 * @param	line	line number of the profile block start
 		 */
-		Profile(int32_t frame, const char* name, const char* file, int line);
+		Profile(int64_t frame, const char* name, const char* file, int line);
 
 		/**
 		 * Destructor
@@ -81,7 +81,7 @@ namespace griffin {
 	// TEMP
 
 	template <int Th>
-	Profile<Th>::Profile(int32_t frame, const char* name, const char* file, int line) :
+	Profile<Th>::Profile(int64_t frame, const char* name, const char* file, int line) :
 		m_frame{ frame },
 		m_name{ name },
 		m_file{ file },
